requirements.txt
transformers
torch
requests
pillow
fastapi
uvicorn
streamlit
python-multipart